Inline high-confidence unit tests into source files and remove duplicated tests (pagerank, graph, search, regex, trigrams, writer_utils, typesitter, plan_helpers)
